+++ app/controllers/environment_variables_controller.rb
@@ -14,7 +14,7 @@ class EnvironmentVariablesController < ApplicationController
     if @environment_variable.valid?
       CreateEnvironmentVariable.new(infrastructure: @infrastructure)
         .call(environment_variable: @environment_variable)
-      redirect_to infrastructure_path(@infrastructure)
+      redirect_to infrastructure_environment_variables_path(@infrastructure)
     else
       render :new
     end
@@ -27,7 +27,12 @@ class EnvironmentVariablesController < ApplicationController
     result = DeleteEnvironmentVariable.new(infrastructure: @infrastructure).call(environment_variable: environment_variable)
 
     flash_message = result.success? ? {notice: "#{environment_variable.name} has been successfully deleted"} : {error: result.error_message}
-    redirect_to infrastructure_path(@infrastructure), flash: flash_message
+    redirect_to infrastructure_environment_variables_path(@infrastructure), flash: flash_message
+  end
+
+  def index
+    @infrastructure = Infrastructure.find(params[:infrastructure_id])
+    @environment_variables = FindEnvironmentVariables.new(infrastructure: @infrastructure).call
   end
 
   def service_name